+++ recipe/thirdparty/providers/config_utils.go
@@ -28,6 +28,7 @@ func getProviderConfigForClient(config tpmodels.ProviderConfig, clientConfig tpm
 		OIDCDiscoveryEndpoint:            config.OIDCDiscoveryEndpoint,
 		UserInfoMap:                      config.UserInfoMap,
 		ValidateIdTokenPayload:           config.ValidateIdTokenPayload,
+		ValidateAccessToken:              config.ValidateAccessToken,
 		RequireEmail:                     config.RequireEmail,
 		GenerateFakeEmail:                config.GenerateFakeEmail,
 	}
@@ -75,6 +76,8 @@ func createProvider(input tpmodels.ProviderInput) *tpmodels.TypeProvider {
 		return Linkedin(input)
 	} else if strings.HasPrefix(input.Config.ThirdPartyId, "boxy-saml") {
 		return BoxySaml(input)
+	} else if strings.HasPrefix(input.Config.ThirdPartyId, "twitter") {
+		return Twitter(input)
 	}
 
 	return NewProvider(input)